+++ src/kernel_evm/CHANGES.md
@@ -17,6 +17,7 @@
 - Remove `mockup` mode previously used for internal testing only. (!10406)
 - RPCs related to blocks now return default values for the POW fields instead of
   mockup. (!10427)
+- The binary `evm-proxy-server` is renamed to `evm-node`. (!10656)
 
 ### Bug fixes